move auth code to auth mod
